context release bug fix
moved the method for closing windows out of the event loop and into a class method. The idea is that it should be possible to open and close multiple windows before destroying the context. For example (where the bug was discovered) running mandelbrot viewer and a julia set viewer as a sub window - when the julia viewer is closed it should not destroy the context. Only after the mandelbrot viewer is finished should this be done.